# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Swagger OpenApi description",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/planos": {
            "get": {
                "tags": [
                    "Planos"
                ],
                "summary": "Get list of planos",
                "operationId": "f756579110476a0107fbee0941c9449d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nome": {
                                                "type": "string"
                                            },
                                            "desconto": {
                                                "type": "string"
                                            },
                                            "ativo": {
                                                "type": "boolean"
                                            },
                                            "descontoAnual": {
                                                "type": "string"
                                            },
                                            "descontoMensal": {
                                                "type": "string"
                                            },
                                            "preco": {
                                                "type": "string"
                                            },
                                            "descricao": {
                                                "type": "string"
                                            },
                                            "frete": {
                                                "type": "string"
                                            },
                                            "qntCafe": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Planos"
                ],
                "summary": "Create a new plano",
                "operationId": "a8536914317c7596cf79d2dc178cc2e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "desconto": {
                                        "type": "string"
                                    },
                                    "ativo": {
                                        "type": "boolean"
                                    },
                                    "descontoAnual": {
                                        "type": "string"
                                    },
                                    "descontoMensal": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "frete": {
                                        "type": "string"
                                    },
                                    "qntCafe": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plano created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/planos/{id}": {
            "get": {
                "tags": [
                    "Planos"
                ],
                "summary": "Get plano by ID",
                "operationId": "21883f815ec86905bd8bd7306d1d6620",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string"
                                        },
                                        "desconto": {
                                            "type": "string"
                                        },
                                        "ativo": {
                                            "type": "boolean"
                                        },
                                        "descontoAnual": {
                                            "type": "string"
                                        },
                                        "descontoMensal": {
                                            "type": "string"
                                        },
                                        "preco": {
                                            "type": "string"
                                        },
                                        "descricao": {
                                            "type": "string"
                                        },
                                        "frete": {
                                            "type": "string"
                                        },
                                        "qntCafe": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plano not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Planos"
                ],
                "summary": "Update a plano",
                "operationId": "42cc1a4bdc6a810715fdc8dafa47e19e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "desconto": {
                                        "type": "string"
                                    },
                                    "ativo": {
                                        "type": "boolean"
                                    },
                                    "descontoAnual": {
                                        "type": "string"
                                    },
                                    "descontoMensal": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "frete": {
                                        "type": "string"
                                    },
                                    "qntCafe": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plano updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string"
                                        },
                                        "desconto": {
                                            "type": "string"
                                        },
                                        "ativo": {
                                            "type": "boolean"
                                        },
                                        "descontoAnual": {
                                            "type": "string"
                                        },
                                        "descontoMensal": {
                                            "type": "string"
                                        },
                                        "preco": {
                                            "type": "string"
                                        },
                                        "descricao": {
                                            "type": "string"
                                        },
                                        "frete": {
                                            "type": "string"
                                        },
                                        "qntCafe": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plano not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planos"
                ],
                "summary": "Delete a plano",
                "operationId": "23fd654c318c8e11d26c3d4610ce468f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plano deleted successfully"
                    },
                    "404": {
                        "description": "Plano not found"
                    }
                }
            }
        },
...
    },
...
}